refactor(backend): renaming route enums, validation-schemas ss-268
