+++ packages/shared/src/modules/routes/libs/enums/enums.ts
@@ -1,3 +1,3 @@
-export { RoutesApiPath } from "./routes-api-path.enum.js";
-export { RoutesValidationMessage } from "./routes-validation-message.enum.js";
-export { RoutesValidationRule } from "./routes-validation-rule.enum.js";
+export { RoutesApiPath } from "./route-api-path.enum.js";
+export { RoutesValidationMessage } from "./route-validation-message.enum.js";
+export { RoutesValidationRule } from "./route-validation-rule.enum.js";
