+++ packages/shared/src/modules/routes/libs/enums/route-validation-message.enum.ts
@@ -1,4 +1,4 @@
-import { RoutesValidationRule } from "./routes-validation-rule.enum.js";
+import { RoutesValidationRule } from "./route-validation-rule.enum.js";
 
 const RoutesValidationMessage = {
 	DESCRIPTION_MAXIMUM_LENGTH: `Description more than ${String(RoutesValidationRule.NAME_MAXIMUM_LENGTH)} characters.`,
