+++ packages/shared/src/modules/routes/libs/validation-schemas/validation-schemas.ts
@@ -1,4 +1,4 @@
-export { routesConstruct } from "./routes-construct.validation-schema.js";
-export { routesCreate } from "./routes-create.validation-schema.js";
-export { routesSearchQuery } from "./routes-search-query.validation-schema.js";
-export { routesUpdate } from "./routes-update.validation-schema.js";
+export { routesConstruct } from "./route-construct.validation-schema.js";
+export { routesCreate } from "./route-create.validation-schema.js";
+export { routesSearchQuery } from "./route-search-query.validation-schema.js";
+export { routesUpdate } from "./route-update.validation-schema.js";
